[Encoding][NFC] Move Encoding Utils out from IR definition.
