+++ compiler/src/iree/compiler/Codegen/Common/CPU/CMakeLists.txt
@@ -84,6 +84,7 @@ iree_cc_library(
     iree::compiler::Codegen::Transforms
     iree::compiler::Codegen::Utils
     iree::compiler::Dialect::Encoding::IR
+    iree::compiler::Dialect::Encoding::Utils
     iree::compiler::Dialect::HAL::IR
   PUBLIC
 )
