+++ compiler/src/iree/compiler/Codegen/ExternalInterfaces/CMakeLists.txt
@@ -38,6 +38,7 @@ iree_cc_library(
     iree::compiler::Codegen::Dialect::GPU::TargetUtils::KnownTargets
     iree::compiler::Codegen::Utils
     iree::compiler::Dialect::Encoding::IR
+    iree::compiler::Dialect::Encoding::Utils
     iree::compiler::Dialect::TensorExt::IR
   PUBLIC
 )
