+++ compiler/src/iree/compiler/Dialect/Encoding/IR/EncodingTypes.h
@@ -19,6 +19,42 @@
 
 namespace mlir::iree_compiler::IREE::Encoding {
 
+static constexpr char kEncodingResolverAttrName[] = "iree.encoding.resolver";
+
+/// Assign a name to operand indices for clarity
+const int64_t MATMUL_LHS = 0;
+const int64_t MATMUL_RHS = 1;
+const int64_t MATMUL_RESULT = 2;
+
+/// Convert operand index to strings for printing
+std::string stringifyOperandIndex(IntegerAttr);
+
+/// Designates a dimension in a matmul (either the M or the N dimension) as
+/// being "narrow", i.e. small enough that we bother lowering the amount of
+/// padding along that dimension compared to how padding we apply to
+/// sufficiently large dimensions.
+struct MatmulNarrowDim {
+  // Enumerates dimensions of a matmul that may be labelled as narrow.
+  enum class Dim {
+    None,
+    M,
+    N,
+  };
+  Dim dim = Dim::None; // Which dimension is designated by *this.
+  int64_t size = 0;    // Size of the designated dimension, or kDynamic.
+
+  explicit operator bool() const { return dim != Dim::None; }
+  bool isM() const { return dim == Dim::M; }
+  bool isN() const { return dim == Dim::N; }
+};
+
+/// Returns the narrow dim in a given `linalgOp`, with respect to the given
+/// `narrowThreshold` below which a dimension is eligible to be considered
+/// narrow. If both M and N are narrow, M is returned. If neither M nor N are
+/// narrow, this returns a default-constructed falsish value.
+MatmulNarrowDim getMatmulNarrowDim(linalg::LinalgOp linalgOp,
+                                   int narrowThreshold);
+
 // The structs defined here because they are used by encoding_interfaces.td.
 struct PropagationEncoding {
   SmallVector<Attribute> operandEncodings;
@@ -56,79 +92,5 @@ struct PropagationResult {
 #include "iree/compiler/Dialect/Encoding/IR/EncodingTypeInterfaces.h.inc" // IWYU pragma: export
 #include "iree/compiler/Dialect/Encoding/IR/EncodingOpInterfaces.h.inc" // IWYU pragma: export
 // clang-format on
-
-//===---------------------------------------------------------------------===//
-// Encoding Dialect Helpers
-//===---------------------------------------------------------------------===//
-
-namespace mlir::iree_compiler::IREE::Encoding {
-
-static constexpr char kEncodingResolverAttrName[] = "iree.encoding.resolver";
-
-/// Returns the encoding attribute from the type if there is an encoding that
-/// implements SerializableEncodingAttrInterface. Otherwise, returns null.
-SerializableEncodingAttrInterface
-getSerializableEncodingAttrInterface(RankedTensorType type);
-
-/// Returns the encoding attribute from the type if there is an encoding.
-/// Otherwise, returns null.
-EncodingAttr getEncodingAttr(RankedTensorType type);
-
-/// Returns true if the type contains packed_storage attribute.
-bool hasPackedStorageAttr(RankedTensorType type);
-
-/// Returns the ContractionDimensions for the encoding user_indexing_maps.
-FailureOr<linalg::ContractionDimensions>
-getEncodingContractionDims(EncodingAttr encoding);
-
-/// Assign a name to operand indices for clarity
-const int64_t MATMUL_LHS = 0;
-const int64_t MATMUL_RHS = 1;
-const int64_t MATMUL_RESULT = 2;
-
-/// Convert operand index to strings for printing
-std::string stringifyOperandIndex(IntegerAttr);
-
-/// Designates a dimension in a matmul (either the M or the N dimension) as
-/// being "narrow", i.e. small enough that we bother lowering the amount of
-/// padding along that dimension compared to how padding we apply to
-/// sufficiently large dimensions.
-struct MatmulNarrowDim {
-  // Enumerates dimensions of a matmul that may be labelled as narrow.
-  enum class Dim {
-    None,
-    M,
-    N,
-  };
-  Dim dim = Dim::None; // Which dimension is designated by *this.
-  int64_t size = 0;    // Size of the designated dimension, or kDynamic.
-
-  explicit operator bool() const { return dim != Dim::None; }
-  bool isM() const { return dim == Dim::M; }
-  bool isN() const { return dim == Dim::N; }
-};
-
-/// Returns the narrow dim in a given `linalgOp`, with respect to the given
-/// `narrowThreshold` below which a dimension is eligible to be considered
-/// narrow. If both M and N are narrow, M is returned. If neither M nor N are
-/// narrow, this returns a default-constructed falsish value.
-MatmulNarrowDim getMatmulNarrowDim(linalg::LinalgOp linalgOp,
-                                   int narrowThreshold);
-
-/// Returns the narrow dim in a given `encoding`, ceiled to a power of two. This
-/// works by inspecting the `iteration_sizes` array attribute in the `encoding`.
-/// If the `iteration_sizes` of one dimension (M or N) is smaller than the
-/// other, then that's the narrow dimension, because the only way it would have
-/// been set to be smaller in the first place, is if we previously flagged that
-/// dimension as narrow. If the `iteration_sizes` of the M and N dimensions
-/// agree, then neither is a narrow dimension and this returns a
-/// default-constructed falsish value.
-MatmulNarrowDim getPo2MatmulNarrowDim(EncodingAttr encoding);
-
-/// Returns true if `encoding` represents a narrow-N matmul RESULT, e.g. the
-/// result of a matvec.
-bool isNarrowNResult(EncodingAttr encoding);
-
-} // namespace mlir::iree_compiler::IREE::Encoding
 
 #endif // IREE_COMPILER_DIALECT_ENCODING_IR_ENCODINGTYPES_H_
