+++ compiler/src/iree/compiler/Dialect/Encoding/Utils/CMakeLists.txt
@@ -1,6 +1,6 @@
 ################################################################################
 # Autogenerated by build_tools/bazel_to_cmake/bazel_to_cmake.py from           #
-# compiler/src/iree/compiler/Dialect/Stream/Conversion/HALToStream/BUILD.bazel #
+# compiler/src/iree/compiler/Dialect/Encoding/Utils/BUILD.bazel                #
 #                                                                              #
 # Use iree_cmake_extra_content from iree/build_defs.oss.bzl to add arbitrary   #
 # CMake-only content.                                                          #
@@ -12,22 +12,17 @@ iree_add_all_subdirs()
 
 iree_cc_library(
   NAME
-    HALToStream
+    Utils
   HDRS
-    "Patterns.h"
+    "Utils.h"
   SRCS
-    "Patterns.cpp"
+    "Utils.cpp"
   DEPS
-    MLIRArithDialect
-    MLIRFuncDialect
-    MLIRFunctionInterfaces
     MLIRIR
-    MLIRTransformUtils
-    MLIRTransforms
-    iree::compiler::Dialect::Encoding::Utils
-    iree::compiler::Dialect::HAL::IR
-    iree::compiler::Dialect::Stream::Conversion
-    iree::compiler::Dialect::Stream::IR
+    MLIRInferTypeOpInterface
+    MLIRLinalgDialect
+    MLIRLinalgUtils
+    iree::compiler::Dialect::Encoding::IR
   PUBLIC
 )
