+++ compiler/src/iree/compiler/Dialect/Stream/Conversion/HALToStream/BUILD.bazel
@@ -21,7 +21,7 @@ iree_compiler_cc_library(
         "Patterns.h",
     ],
     deps = [
-        "//compiler/src/iree/compiler/Dialect/Encoding/IR",
+        "//compiler/src/iree/compiler/Dialect/Encoding/Utils",
         "//compiler/src/iree/compiler/Dialect/HAL/IR",
         "//compiler/src/iree/compiler/Dialect/Stream/Conversion",
         "//compiler/src/iree/compiler/Dialect/Stream/IR",
