+++ compiler/src/iree/compiler/Dialect/Stream/IR/CMakeLists.txt
@@ -54,7 +54,7 @@ iree_cc_library(
     MLIRTensorDialect
     MLIRTransformUtils
     MLIRViewLikeInterface
-    iree::compiler::Dialect::Encoding::IR
+    iree::compiler::Dialect::Encoding::Utils
     iree::compiler::Dialect::Util::IR
     iree::compiler::Utils
   PUBLIC
