+++ compiler/src/iree/compiler/Dialect/Stream/IR/StreamOpFolders.cpp
@@ -7,7 +7,7 @@
 #include <algorithm>
 #include <optional>
 
-#include "iree/compiler/Dialect/Encoding/IR/EncodingTypes.h"
+#include "iree/compiler/Dialect/Encoding/Utils/Utils.h"
 #include "iree/compiler/Dialect/Stream/IR/StreamOps.h"
 #include "iree/compiler/Dialect/Util/IR/ClosureOpUtils.h"
 #include "iree/compiler/Dialect/Util/IR/UtilTypes.h"
