+++ compiler/src/iree/compiler/Utils/CMakeLists.txt
@@ -62,6 +62,7 @@ iree_cc_library(
     iree::base::internal::flatcc::debugging
     iree::base::internal::flatcc::parsing
     iree::compiler::Dialect::Encoding::IR
+    iree::compiler::Dialect::Encoding::Utils
     iree::compiler::Dialect::Util::IR
   PUBLIC
 )
